refactor: @EnableJpaAuditing 위치 이동
- 한 곳에서 설정을 관리하기 위해 Caffeine 설정 클래스로 어노테이션 이동
