+++ order/src/main/java/project/swithme/order/SwithMeOrderApplication.java
@@ -2,10 +2,8 @@ package project.swithme.order;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.cache.annotation.EnableCaching;
 import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
 
-@EnableCaching
 @EnableJpaAuditing
 @SpringBootApplication
 public class SwithMeOrderApplication {
