+++ order/src/main/java/project/swithme/order/common/configuration/caffeine/CaffeineConfiguration.java
@@ -1,13 +1,15 @@
-package project.study.caffeine.common.configuration;
+package project.swithme.order.common.configuration.caffeine;
 
 import com.github.benmanes.caffeine.cache.Caffeine;
 import org.springframework.cache.CacheManager;
+import org.springframework.cache.annotation.EnableCaching;
 import org.springframework.cache.caffeine.CaffeineCacheManager;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
 import static java.util.concurrent.TimeUnit.DAYS;
 
+@EnableCaching
 @Configuration
 public class CaffeineConfiguration {
